requirements.txt
attrs==21.2.0
aws-cdk.assertions==1.140.0
aws-cdk.aws-ec2==1.140.0
aws-cdk.aws-glue==1.140.0
aws-cdk.aws-iam==1.140.0
aws-cdk.aws-kinesis==1.140.0
aws-cdk.aws-lambda==1.140.0
aws-cdk.aws-s3==1.140.0
aws-cdk.aws-apigateway==1.140.0
aws-cdk.aws-kinesisfirehose==1.140.0
aws-cdk.aws-route53==1.140.0
aws-cdk.core==1.140.0
boto3==1.18.61
botocore==1.21.61
chevron==0.14.0
click==8.0.3
jmespath==0.10.0
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pylint==2.12.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
six==1.15.0
toml==0.10.2
tox==3.24.4
typing-extensions~=3.7
urllib3==1.26.6
utils==1.0.1
